refactor(qwen35): move scheduler tests into module
